// old/src/app/core/osmose-api-client/model/mailingListOutput.ts
/**
 * osmose-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MembreOuput } from './membreOuput';
import { CentreOutput } from './centreOutput';
import { TypesMembres } from './typesMembres';


export interface MailingListOutput { 
    id?: number | null;
    libelle?: string | null;
    typeMembre?: TypesMembres;
    readonly listePersonnalisee?: boolean;
    centre?: CentreOutput;
    membres?: Array<MembreOuput> | null;
}
export namespace MailingListOutput {
}